ml_model/requirements.txt
numpy==1.24.3
Pillow==10.1.0
tensorflow==2.15.0

